added newPost component which is responsible for adding picture with description into categories
